Fix AutonomousDecisionEngine decimal precision issues (8 locations)
Per COMPLETE_FIX_GUIDE Week 1 priorities:
- Changed CalculateRSI return type: double → decimal (removed cast, use decimal literals)
- Changed CalculateMACD return type: double → decimal (removed cast)
- Changed CalculateBollingerPosition return type: double → decimal (use decimal literals)
- Changed CalculateATR return type: double → decimal (removed cast)
- Changed CalculateVolumeMA return type: double → decimal (removed cast)
- Changed CalculateTechnicalIndicatorsAsync return type: Dictionary<string, double> → Dictionary<string, decimal>
- Updated indicators dictionary initialization with decimal literals (m suffix)
- Changed currentPrice and currentVolume to decimal types
- Updated fallback price calculations to use decimal literals
- Changed MarketContext.TechnicalIndicators: Dictionary<string, double> → Dictionary<string, decimal> (both in BotCore.Brain.Models and TradingBot.Abstractions)
- Updated MarketContext.Price and Volume to decimal in TradingBot.Abstractions
- Fixed NeuralUcbExtended.cs ATR volatility calculation to use decimal Math.Min

All indicator calculations now maintain full decimal precision without unnecessary double conversions.

Note: This creates cascading type errors in IntelligenceStack that need to be fixed next.

+++ src/Abstractions/TradingInterfaces.cs
@@ -33,11 +33,11 @@ public class MarketContext
 {
     public string Symbol { get; set; } = string.Empty;
     public DateTime Timestamp { get; set; } = DateTime.UtcNow;
-    public double Price { get; set; }
-    public double Volume { get; set; }
+    public decimal Price { get; set; }
+    public decimal Volume { get; set; }
     public double Bid { get; set; }
     public double Ask { get; set; }
-    public Dictionary<string, double> TechnicalIndicators { get; } = new();
+    public Dictionary<string, decimal> TechnicalIndicators { get; } = new();
     public RegimeState? CurrentRegime { get; set; }
     
     // Additional properties for intelligence integration
